+++ GasNetSim/components/utils/gas_mixture/GERG2008/gerg2008.py
@@ -86,6 +86,7 @@
 # static void PseudoCriticalPointGERG(const std::vector<double> &x, double &Tcx, double &Dcx);
 # static void tTermsGERG(const double lntau, const std::vector<double> &x);
 # """
+from collections import OrderedDict
 
 #   #!/usr/bin/env python
 #   -*- coding: utf-8 -*-
@@ -98,6 +99,9 @@
 
 import numpy as np
 import math
+from collections import Counter
+from .setup import *
+from copy import deepcopy
 
 from GasNetSim.components.utils.gas_mixture.GERG2008.setup import *
 
@@ -141,17 +145,14 @@ The compositions in the x() array use the following order and must be sent as mo
 
 
 class GasMixtureGERG2008:
-    def __init__(self, P, T, composition, pressure_unit='Pa', temperature_unit='K'):
+    def __init__(self, P_Pa, T_K, composition):
         # Input parameters
-        if pressure_unit == 'Pa':
-            self.P = P / 1000  # Pa -> kPa
-        elif pressure_unit == 'kPa':
-            self.P = P
-        elif pressure_unit == 'bar':
-            self.P = 101.325 * P  # bar -> kPa
-        else:
-            raise ValueError("Pressure unit is not known!")
-        self.T = T
+        self.dPdT = None
+        self.d2PdD2 = None
+        self.dPdD = None
+        self.P = P_Pa / 1000  # Pa -> kPa
+
+        self.T = T_K
         self.x = self.CovertCompositionGERG(composition=composition)  # gas composition
 
         # Calculated properties
@@ -169,10 +170,166 @@ class GasMixtureGERG2008:
         self.c = 0  # speed of sound [m/s]
         self.gibbs_energy = 0  # Gibbs energy [J/mol]
 
-        self.JT = 1  # Joule-Thomson coefficient [K/kPa]
+        self.JT = 1  # Joule-Thomson coefficient [K/Pa]
         self.isentropic_exponent = 0  # Isentropic exponent
 
+        self.R_specific = 0
+        self.viscosity = 2e-4  # TODO add function
+
         self.PropertiesGERG()
+
+    def CalculateHeatingValue(self, comp, hhv, parameter):
+
+        dict_components = {'methane': {'C': 1, 'H': 4},
+                           'nitrogen': {'N': 2},
+                           'carbon dioxide': {'C': 1, 'O': 2},
+                           'ethane': {'C': 2, 'H': 6},
+                           'propane': {'C': 3, 'H': 8},
+                           'isobutane': {'C': 4, 'H': 10},
+                           'n-butane': {'C': 4, 'H': 10},
+                           'isopentane': {'C': 5, 'H': 12},
+                           'n-pentane': {'C': 5, 'H': 12},
+                           'n-hexane': {'C': 6, 'H': 14},
+                           'n-heptane': {'C': 7, 'H': 16},
+                           'n-octane': {'C': 8, 'H': 18},
+                           'n-nonane': {'C': 9, 'H': 20},
+                           'n-decane': {'C': 10, 'H': 22},
+                           'hydrogen': {'H': 2},
+                           'oxygen': {'O': 2},
+                           'carbon monoxide': {'C': 1, 'O': 1},
+                           'water': {'H': 2, 'O': 1},
+                           'hydrogen sulfide': {'H': 2, 'S': 1},
+                           'helium': {'He': 1},
+                           'argon': {'Ar': 1}}
+                           # 'SO2': {'S': 1, 'O': 2}}
+        # 298 K
+        # dict_enthalpy_mole = {'methane': -74602.416533355,
+        #                       'nitrogen': 0.0,
+        #                       'carbon dioxide': -393517.79827154,
+        #                       'ethane': -83856.2627150042,
+        #                       'propane': -103861.117481869,
+        #                       'isobutane': -135360.0,
+        #                       'n-butane': -125849.99999999999,
+        #                       'isopentane': -178400.0,
+        #                       'n-pentane': -173500.0,
+        #                       'n-hexane': -198490.0,
+        #                       'n-heptane': -223910.0,
+        #                       'n-hctane': -249730.0,
+        #                       'n-nonane': -274700.0,
+        #                       'n-decane': -300900.0,
+        #                       'hydrogen': 0.0,
+        #                       'oxygen': -4.40676212751828,
+        #                       'carbon monoxide': -110525.0,
+        #                       'water': -241833.418361837,
+        #                       'hydrogen sulfide': -20600.0,
+        #                       'helium': 0.0,
+        #                       'argon': 0.0,
+        #                       'carbon': 0.0}
+        #                       # 'H': 218000.0,
+        #                       # 'O': 249190.0,
+        #                       # 'SO2': -296840.0}
+
+        # 273 K
+        dict_enthalpy_mole = {'methane': -75483.51423273719,
+                              'nitrogen': 0.0,
+                              'carbon dioxide': -394431.82606764464,
+                              'ethane': -83856.2627150042,
+                              'propane': -103861.117481869,
+                              'isobutane': -135360.0,
+                              'n-butane': -125849.99999999999,
+                              'isopentane': -178400.0,
+                              'n-pentane': -173500.0,
+                              'n-hexane': -198490.0,
+                              'n-heptane': -223910.0,
+                              'n-hctane': -249730.0,
+                              'n-nonane': -274700.0,
+                              'n-decane': -300900.0,
+                              'hydrogen': 0.0,
+                              'oxygen': -4.40676212751828,
+                              'carbon monoxide': -111262.34509634285,
+                              'water': -242671.7203547155,
+                              'hydrogen sulfide': -20600.0,
+                              'helium': 0.0,
+                              'argon': 0.0,
+                              'carbon': 0.0}
+                              # 'H': 218000.0,
+                              # 'O': 249190.0,
+                              # 'SO2': -296840.0}
+
+        # # reactants
+        # atom_list = []
+        # new_dict = {}
+        #
+        # for key, value in comp.items():
+        #     for key1, value1 in dict_components.items():
+        #         if key == key1:
+        #             new_dict = dict_components.get(key1, {})
+        #             new_dict.update((x, y * value) for x, y in new_dict.items())
+        #             atom_list.append(new_dict)
+        #
+        # # counter_reactants = Counter()
+        # reactants_dict = Counter()
+        # for d in atom_list:
+        #     reactants_dict.update(d)
+
+        # reactants
+        atom_list = []
+
+        for key, value in comp.items():
+            for key1, value1 in dict_components.items():
+                if key == key1:
+                    atom_list.append(dict((x, y * value) for x, y in dict_components.get(key1, {}).items()))
+
+        reactants_atom = Counter()
+        for d in atom_list:
+            reactants_atom.update(d)
+
+        # products
+        n_CO2 = reactants_atom["C"]
+        n_SO2 = reactants_atom["S"]
+        n_H2O = reactants_atom["H"] / 2
+        products_dict = {'carbon dioxide': n_CO2, 'sulfur dioxide': n_SO2, 'water': n_H2O}
+
+        # oxygen for complete combustion
+        n_O = n_CO2 * 2 + n_SO2 * 2 + n_H2O * 1  # 2 is number of O atoms in CO2 AND SO2 and 1 is number of O atoms in H2O
+        n_O2 = n_O / 2
+        reactants_dict = deepcopy(comp)
+        reactants_dict.update({'oxygen': n_O2})
+
+        # LHV calculation
+        LHV = 0
+        for key, value in dict_enthalpy_mole.items():
+            for key1, value1 in reactants_dict.items():
+                if key == key1:
+                    LHV += value * value1
+            for key2, value2 in products_dict.items():
+                if key == key2:
+                    LHV -= value * value2
+
+        # 298 K
+        hw_liq = -285825.0
+        hw_gas = -241820.0
+
+        # 273 K
+        # hw_liq = -287654.96084928664
+        # hw_gas = -242628.01574091613
+
+        HHV = LHV + (hw_gas - hw_liq) * products_dict["water"]
+
+        if parameter == 'mass':
+            # returns heating value in kJ/kg
+            if hhv:
+                heating_value = HHV / self.MolarMass
+            else:
+                heating_value = LHV / self.MolarMass
+        else:
+            # returns heating value in kJ/m3
+            if hhv:
+                heating_value = HHV * self.MolarDensity
+            else:
+                heating_value = LHV * self.MolarDensity
+
+        return heating_value
 
     @staticmethod
     def CovertCompositionGERG(composition):
@@ -362,7 +519,7 @@ class GasMixtureGERG2008:
                     Cp - Isobaric heat capacity [J/(mol-K)]
                      W - Speed of sound (m/s)
                      G - Gibbs energy (J/mol)
-                    JT - Joule-Thomson coefficient (K/kPa)
+                    JT - Joule-Thomson coefficient (K/Pa)
                  Kappa - Isentropic Exponent
                      A - Helmholtz energy (J/mol)
         """
@@ -425,10 +582,11 @@ class GasMixtureGERG2008:
         self.Cp = Cp
         self.c = W  # speed of sound [m/s]
         self.gibbs_energy = G  # Gibbs energy [J/mol]
-        self.JT = JT  # Joule-Thomson coefficient [K/kPa]
+        self.JT = JT / 1e3  # Joule-Thomson coefficient [K/Pa]
         self.isentropic_exponent = Kappa  # Isentropic exponent
-        self.rho = 0
-        self.SG = 1
+        self.rho = self.MolarMass * self.P * 1000 / RT
+        self.SG = self.MolarMass / air_molar_mass
+        self.R_specific = R / self.MolarMass
 
     def Alpha0GERG(self):
         """